+++ src/styles/Button.css
@@ -1,4 +1,4 @@
-.storybook-button {
+.button {
   font-family: 'Nunito Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
   font-weight: 700;
   border: 0;
@@ -7,24 +7,25 @@
   display: inline-block;
   line-height: 1;
 }
-.storybook-button--primary {
-  color: white;
-  background-color: #1ea7fd;
-}
-.storybook-button--secondary {
-  color: #333;
-  background-color: transparent;
-  box-shadow: rgba(0, 0, 0, 0.15) 0px 0px 0px 1px inset;
-}
-.storybook-button--small {
+.button--small {
   font-size: 12px;
   padding: 10px 16px;
 }
-.storybook-button--medium {
+.button--medium {
   font-size: 14px;
   padding: 11px 20px;
 }
-.storybook-button--large {
+.button--large {
   font-size: 16px;
   padding: 12px 24px;
+}
+
+.button__icon {
+  position: relative;
+  top: 1px;
+  display: inline-block;
+  width: 20px;
+  height: 20px;
+  margin-right: 8px;
+  vertical-align: text-top;
 }
